Move preview proxy logic out to own singleton
