+++ app/com/gu/viewer/controllers/Proxy.scala
@@ -1,45 +1,13 @@
 package com.gu.viewer.controllers
 
 import com.gu.viewer.config.Configuration
-import com.gu.viewer.proxy.{Proxy => NewProxy}
-import com.gu.viewer.views.html
-import java.net.URLEncoder
+import com.gu.viewer.proxy.{Proxy => NewProxy, PreviewProxy}
 import javax.inject.Inject
 import com.gu.viewer.logging.Loggable
-import play.api.libs.concurrent.Execution.Implicits._
-import play.api.Play.current
-import play.api.libs.ws.{WSResponse, WSClient, WSRequest}
-import play.api.mvc.{Controller, Cookie, Cookies, Action, Result, RequestHeader}
-import scala.concurrent.Future
+import play.api.mvc.{Controller, Action}
 
 
-class Proxy @Inject() (newProxy: NewProxy) extends Controller with Loggable {
-
-  private val COOKIE_PREVIEW_SESSION = "PLAY_SESSION"
-  private val COOKIE_PREVIEW_AUTH = "GU_PV_AUTH"
-
-  private val SESSION_KEY_PREVIEW_SESSION = "preview-session"
-  private val SESSION_KEY_PREVIEW_AUTH = "preview-auth"
-  private val SESSION_KEY_RETURN_URL = "preview-auth-return-url"
-
-  val previewLoginUrl = s"http://${Configuration.previewHost}/login"
-
-  def loginCallbackUrl(implicit request: RequestHeader) = {
-    val protocol = if (request.secure) "https" else "http"
-    s"$protocol://${request.host}${routes.Proxy.previewAuthCallback()}"
-  }
-
-  /**
-   * Transform proxy server relative URI to viewer URI.
-   */
-  def proxyUriToViewerUri(uri: String) = {
-    val proxyUri = """^\/proxy(\/.+)$""".r
-
-    uri match {
-      case proxyUri(path) => Some(path)
-      case _ => None
-    }
-  }
+class Proxy @Inject() (newProxy: NewProxy, previewProxy: PreviewProxy) extends Controller with Loggable {
 
   def proxy(service: String, path: String) = Action.async { implicit request =>
     val protocol = if (request.secure) "https" else "http"
@@ -49,79 +17,15 @@ class Proxy @Inject() (newProxy: NewProxy) extends Controller with Loggable {
     }
     val url = s"$protocol://$serviceHost/$path"
 
+    if (service == "preview") {
+      previewProxy.proxy(request.secure, request.host, request.uri, url, request.session)
 
-    def doPreviewAuth() = {
-      val proxyRequestUrl = previewLoginUrl
-
-      log.info(s"Proxy Preview auth to: $proxyRequestUrl")
-
-      def handleResponse(response: WSResponse) = {
-
-        // store new preview session from response
-        val cookies = Cookies.fromSetCookieHeader(response.header("Set-Cookie"))
-        val previewSessionOpt = cookies.get(COOKIE_PREVIEW_SESSION).map( c => SESSION_KEY_PREVIEW_SESSION -> c.value )
-
-        val loc = response.header("Location").get
-
-        previewSessionOpt match {
-          case Some(session) => {
-            val returnUrl = SESSION_KEY_RETURN_URL -> proxyUriToViewerUri(request.uri).getOrElse(request.uri)
-            Ok(html.loginRedirect(loc))
-              .withSession(request.session - SESSION_KEY_PREVIEW_SESSION - SESSION_KEY_PREVIEW_AUTH + session + returnUrl)
-          }
-
-          case None => badGatewayResponse("Unexpected response from preview login request", response)
-        }
-      }
-
-      newProxy.post(proxyRequestUrl, queryString = Seq("redirect-url" -> loginCallbackUrl)) {
-        case response if response.status == 303 => Future.successful(handleResponse(response))
-        // TODO should we handle non redirect responses here?
-      }
-    }
-
-
-    def doPreviewProxy(session: String, auth: String) = {
-      log.info(s"Proxy to preview: $url")
-
-      val cookies = Seq(COOKIE_PREVIEW_SESSION -> session, COOKIE_PREVIEW_AUTH -> auth).map(c => Cookie(c._1, c._2))
-
-      /* TODO handle redirects with proxy
-          case (status, Some(otherLocation)) => {
-            log.warn(s"Proxied response for $url is $status redirect to: $otherLocation")
-            Future.successful(Status(status).withHeaders("Location" -> otherLocation))
-          }
-      */
-
-      def isLoginRedirect(response: WSResponse) = {
-        response.status == 303 &&
-        response.header("Location").exists(l => l == previewLoginUrl || l == "/login")
-      }
-
-      newProxy.get(url, cookies = cookies) {
-        case response if isLoginRedirect(response) => doPreviewAuth()
-      }
-
-    }
-
-
-    def doProxy() = {
+    } else {
+      // live
       log.info(s"Proxy to: $url")
 
       // TODO rewrite redirects to proxied URLS
       newProxy.get(url, followRedirects = true)()
-    }
-
-
-    if (service == "preview") {
-      request.session.get(SESSION_KEY_PREVIEW_SESSION) -> request.session.get(SESSION_KEY_PREVIEW_AUTH) match {
-        case (Some(session), Some(auth)) => doPreviewProxy(session, auth)
-        case _ => doPreviewAuth()
-      }
-
-    } else {
-      // live
-      doProxy()
     }
 
   }
@@ -133,54 +37,8 @@ class Proxy @Inject() (newProxy: NewProxy) extends Controller with Loggable {
    * Proxy all request params and Preview session cookie to Preview authentication callback.
    * Store response cookies into Viewer's play session.
    */
-  def previewAuthCallback = Action.async { implicit request =>
-
-    val queryParams = request.queryString.flatMap( q => q._2.map { v => q._1 -> v } ).toSeq :+ ("redirect-url" -> loginCallbackUrl)
-
-    def handleResponse(response: WSResponse): Result = {
-      val newCookies: Map[String, Cookie] = response.allHeaders.get("Set-Cookie") match {
-        case Some(setCookieHeader) => setCookieHeader
-          .flatMap { h => Cookies.fromSetCookieHeader(Some(h)) }
-          .groupBy(_.name)
-          .mapValues(_.head)
-
-        case None => Map.empty
-      }
-
-      val responseCookies = Cookies.fromCookieHeader(response.header("Cookie"))
-
-      val sessionOpt = newCookies.get(COOKIE_PREVIEW_SESSION)
-        .orElse(responseCookies.get(COOKIE_PREVIEW_SESSION))
-        .map(c => SESSION_KEY_PREVIEW_SESSION -> c.value)
-
-      val authOpt = newCookies.get(COOKIE_PREVIEW_AUTH).map(c => SESSION_KEY_PREVIEW_AUTH -> c.value)
-
-      (sessionOpt, authOpt) match {
-        case (Some(sessionValue), Some(authValue)) => {
-          val returnUrl = request.session.get(SESSION_KEY_RETURN_URL).getOrElse("/proxy/preview/uk")
-          Redirect(returnUrl)
-            .withSession(request.session - SESSION_KEY_RETURN_URL + sessionValue + authValue)
-        }
-        case (None, None) => badGatewayResponse("Bad response from preview auth callback", response)
-        case (None, _) => badGatewayResponse("Preview Session cookie not returned", response)
-        case (_, None) => badGatewayResponse("Preview Auth cookie not returned", response)
-      }
-    }
-
-    request.session.get(SESSION_KEY_PREVIEW_SESSION) match {
-      case None => Future.successful(BadRequest("Preview session not established"))
-
-      case Some(session) => {
-        val proxyUrl = s"http://${Configuration.previewHost}/oauth2callback"
-        log.info(s"Proxy preview auth callback to: $proxyUrl")
-
-        val cookies = Seq(Cookie(COOKIE_PREVIEW_SESSION, session))
-
-        newProxy.get(proxyUrl, queryString = queryParams, cookies = cookies) {
-          case r => Future.successful(handleResponse(r))
-        }
-      }
-    }
+  def previewAuthCallback = Action.async { request =>
+    previewProxy.previewAuthCallback(request.host, request.secure, request.queryString, request.session)
   }
 
 
@@ -201,8 +59,4 @@ class Proxy @Inject() (newProxy: NewProxy) extends Controller with Loggable {
     }
   }
 
-  private def badGatewayResponse(msg: String, response: WSResponse) = {
-    log.warn(s"$msg: ${response.toString} ${response.allHeaders} ${response.body}")
-    BadGateway(msg)
-  }
 }
